+++ R/AllClass.R
@@ -550,6 +550,22 @@ setMethod(f="[[",
                   return(if (missing(j)) swSigma3(x) else swSigma3(x, eos=j))
               } else if (i == "sigma4") {
                   return(if (missing(j)) swSigma4(x) else swSigma4(x, eos=j))
+              } else if (i %in% paste0("spiciness", 0:2)) {
+                  #message("in AllClass")
+                  salinity <- x[["salinity"]]
+                  temperature <- x[["temperature"]]
+                  pressure <- x[["pressure"]]
+                  longitude <- x[["longitude"]]
+                  latitude <- x[["latitude"]]
+                  SA <- gsw::gsw_SA_from_SP(SP=salinity,
+                      p=pressure,
+                      longitude=longitude,
+                      latitude=latitude)
+                  CT <- gsw::gsw_CT_from_t(SA, temperature, pressure)
+                  return(switch(i,
+                          "spiciness0"=gsw::gsw_spiciness0(SA, CT),
+                          "spiciness1"=gsw::gsw_spiciness1(SA, CT),
+                          "spiciness2"=gsw::gsw_spiciness2(SA, CT)))
               } else if (i == "silicate") {
                   return(x@data$silicate)
               } else if (i == paste("sound", "speed")) {